+++ log_merger/__main__.py
@@ -1,3 +1,0 @@
-from log_merger.log_merger import main
-
-main()